+++ Classes/Model/AbstractModel.php
@@ -367,23 +367,6 @@ abstract class AbstractModel extends AbstractObjectWithAccessors implements Iden
     }
 
     /**
-     * Gets the value stored in under the key $key as a collection of models.
-     *
-     * @param string $key the key of the element to retrieve, must not be empty
-     *
-     * @return Collection<AbstractModel> the data item for the given key
-     *
-     * @throws \UnexpectedValueException if there is a data item stored for the key $key that is not a collection
-     *         or if that item has not been set yet
-     *
-     * @deprecated will be removed in oelib 5.0 - use `getAsCollection` instead
-     */
-    public function getAsList(string $key): Collection
-    {
-        return $this->getAsCollection($key);
-    }
-
-    /**
      * Makes sure this model has some data by loading the data for ghost models.
      */
     private function load(): void